requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.14.0
altair==6.0.0
androguard==3.3.5
annotated-types==0.7.0
anyio==4.7.0
apkInspector==1.3.2
appnope==0.1.4
asn1crypto==1.5.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
autopep8==2.1.0
banal==1.0.6
bcrypt==4.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
black==25.1.0
bleach==6.1.0
blinker==1.9.0
blis==1.2.0
cachetools==6.2.4
catalogue==2.0.10
category-encoders==2.6.3
certifi==2024.2.2
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.2.0
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.11
dataset==1.6.2
debugpy==1.8.8
decorator==5.1.1
dnspython==2.6.1
ecdsa==0.19.1
email-validator==2.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.1.0
fastapi==0.115.6
filelock==3.13.3
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.6.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.50.0
frida==16.5.6
frozenlist==1.8.0
fsspec==2024.3.1
future==1.0.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
huggingface-hub==0.27.1
idna==3.6
imageio==2.37.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.29.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kaggle==1.6.8
keras==3.6.0
kiwisolver==1.4.5
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
lime==0.2.0.1
loguru==0.7.2
lxml==5.3.0
Mako==1.3.6
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.7.0
murmurhash==1.0.12
mutf8==1.0.6
mypy_extensions==1.1.0
namex==0.0.8
narwhals==2.15.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt_einsum==3.4.0
optree==0.13.0
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.3.6
pluggy==1.5.0
preshed==3.0.9
prometheus_client==0.21.1
prometheus_flask_exporter==0.23.2
prompt_toolkit==3.0.48
propcache==0.4.1
protobuf==5.26.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.2
pycocotools==2.0.7
pycodestyle==2.11.1
pycparser==3.0
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pydot==3.0.2
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.7.3
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.13.0
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.16.0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.37.0
regex==2024.11.6
requests==2.31.0
requests-file==2.1.0
rich==13.9.4
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
selenium==4.18.1
setuptools==70.2.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.46
srsly==2.5.1
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.2
streamlit==1.53.0
sympy==1.12
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
text-unidecode==1.3
textblob==0.19.0
thinc==8.3.4
threadpoolctl==3.4.0
tifffile==2025.3.30
tldextract==5.1.3
tokenizers==0.21.0
toml==0.10.2
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.64.1
traitlets==5.14.3
transformers==4.48.0
trio==0.25.0
trio-websocket==0.11.1
twilio==9.9.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.34.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.0
wordcloud==1.9.4
wrapt==1.16.0
wsproto==1.2.0
yarl==1.22.0
